+++ Source/_gltf2usd/gltf2/Skin.py
@@ -1,5 +1,5 @@
 from sets import Set
-class Skin:
+class Skin(object):
     """Represents a glTF Skin
     """